Refactor the code: imports, project structure, copyright component
